Extract schema extraction logic and park it in validatePassword.ts
