+++ packages/manager/src/utilities/validatePassword/validatePassword.ts
@@ -1,6 +1,7 @@
 import { string } from 'yup';
 import { MINIMUM_PASSWORD_STRENGTH } from 'src/constants';
 import { PasswordValidationType } from 'src/featureFlags';
+import { object } from 'yup';
 import * as zxcvbn from 'zxcvbn';
 
 const passwordLengthAndCharacterSchema = string()
@@ -30,4 +31,44 @@ export const validatePassword = (
         ? null
         : 'Password does not meet complexity requirements.';
   }
+};
+
+/**
+ * Extends a Yup schema with a password validation
+ * check. This check is dynamic based on the type
+ * of validation requested, normally controlled through
+ * a feature flag.
+ *
+ * Remove this method once API password validation
+ * is stable.
+ *
+ * @todo: Update Yup and typings to avoid the any
+ * dodge here
+ */
+export const extendValidationSchema = (
+  validationType: PasswordValidationType,
+  schema: any
+) => {
+  return (schema as any).concat(
+    object({
+      root_pass: string()
+        .required('Password is required.')
+        .test({
+          name: 'root-password-strength',
+          // eslint-disable-next-line object-shorthand
+          test: function(value) {
+            const passwordError = validatePassword(
+              validationType ?? 'none',
+              value
+            );
+            return Boolean(validatePassword)
+              ? this.createError({
+                  message: passwordError,
+                  path: 'root_pass'
+                })
+              : true;
+          }
+        })
+    })
+  );
 };
